requirements.txt
aspy.yaml==0.3.0
cached-property==1.3.0
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
decorator==4.0.11
Django==1.11.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.0.4
djangorestframework==3.6.3
factory-boy==2.8.1
Faker==0.7.17
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
nodeenv==1.1.2
pexpect==4.2.1
pickleshare==0.7.4
pre-commit==0.14.2
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
redgreenunittest==0.1.1
requests==2.18.1
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.21.1
virtualenv==15.1.0
wcwidth==0.1.7
